Improve VTest

+++ autoload/vtools.vim
@@ -6,7 +6,7 @@ endfunction
 
 function! vtools#fmt() abort
   let l:tmpfile = s:write_to_tmp(expand('%'), '.v')
-  let l:line = line('.')
+  let l:line = line('.'):
   let l:col = col('.')
   echo system('v fmt -w ' . l:tmpfile)
   execute '%!cat ' . l:tmpfile
@@ -14,10 +14,20 @@ function! vtools#fmt() abort
   silent call system('rm ' . l:tmpfile)
 endfunction
 
+
 function! vtools#test() abort
-  let l:tmpfile = s:write_to_tmp(expand('%'), '_test.v')
-  echo system('v test ' . l:tmpfile)
-  silent call system('rm ' . l:tmpfile)
+  let l:name = expand('%')
+  if l:name =~ '_test.v$'
+    let l:tmpfile = s:write_to_tmp(l:name, '_test.v')
+    echo system('v test ' . l:tmpfile)
+    silent call system('rm ' . l:tmpfile)
+  else
+    let l:test_path = fnamemodify(l:name, ':r') . '_test.v'
+    if ! filereadable(l:test_path)
+      let l:test_path = fnamemodify(l:name, ':h')  " fallback to dir
+    endif
+    echo system('v test ' . l:test_path)
+  endif
 endfunction
 
 function! vtools#run() abort
